+++ Ice-cream-Parlour/hello/static/this.txt
@@ -1,0 +1,1 @@
+this is a text in static page.
