+++ Ice-cream-Parlour/hello/template/about.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+{% block title %}About{% endblock title %}
+{% block body %}
+
+{% endblock body %}
