+++ Ice-cream-Parlour/hello/template/contacts.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+{% block title %}Contacts{% endblock title %}
+{% block body %}
+ 
+<div class="container">
+    
+</div>
+
+{% endblock body %}
