+++ Ice-cream-Parlour/hello/template/services.html
@@ -1,0 +1,3 @@
+{% extends "base.html" %}
+{% block title %}Services{% endblock title %}
+{% block body %}This is Services block{% endblock body %}
