--- grocery/AvailableProduct.txt 
@@ -1,13 +1,0 @@
-Item UnitPrice
-Soap 2
-Toothbrush 2.3
-Apple 1.5
-Kiwi 3.4
-Spinach 1.2
-Cookie-Jar 10
-Water-Bottle 2
-Juice 4
-Milk 3
-Coffee 7
-Chips 2
-Eggs 4.5
